+++ src/main/java/com/project/aop/LoginCheckAOP.java
@@ -1,5 +1,7 @@
 package com.project.aop;
 
+import java.util.List;
+
 import javax.servlet.http.HttpSession;
 
 import org.aspectj.lang.JoinPoint;
@@ -11,12 +13,18 @@ import org.aspectj.lang.annotation.Pointcut;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import com.project.dto.MemberDTO;
+import com.project.service.AdminService;
+
 @Component
 @Aspect
 public class LoginCheckAOP {
-
+	
 	@Autowired
 	private HttpSession session;
+	
+	@Autowired
+	private AdminService aservice;
 	
 	@Pointcut("execution(* com.project.controller.BasketController.*(..))")
 	public void basketController() {}
@@ -30,13 +38,15 @@ public class LoginCheckAOP {
 	public void admin() {}
 	
 	
-	@Around("basketController()||shopBoard()||timeLine()||member()||admin()")
+	@Around("basketController()||shopBoard()||timeLine()||member()")
 	public String logCheck(ProceedingJoinPoint pjp) {
-		if(session.getAttribute("id")==null) {
-	
 			
+		if(session.getAttribute("id")==null) {
+		
 			return "loginPlease";
+			
 		}else {
+			
 			try {
 				return (String) pjp.proceed();
 			} catch (Throwable e) {
@@ -45,5 +55,42 @@ public class LoginCheckAOP {
 		return null;
 		}
 	}
+	
+
+	//만일 관리자id로 로그인을 하지 않을 경우 관리자 페이지에 들어갈 수 없게 만든다.
+	@Around("admin()")
+	public String AdminCheck(ProceedingJoinPoint pjp){	
+
+		MemberDTO member = (MemberDTO)session.getAttribute("id");
+
+		//로그인을 하지 않은 경우
+		if(member == null) {
+
+			return "adminPlease";
+		}	
+
+		//로그인은 했지만 관리자 id가 아닌 경우
+		else {
+
+			if(member.getMember_id().equals("admin")) {
+
+				try {
+					return (String) pjp.proceed();
+				} catch (Throwable e) {
+					e.printStackTrace();
+				}
+
+			}else {
+
+				return "adminPlease";
+			}
+		}
+
+		return null;
+	}
+
+
+	
+	
 	
 }
